Crate rustful

Reexports

pub use self::server::Server;

pub use self::context::Context;

pub use self::response::Response;

pub use self::response::Error;

pub use self::handler::Handler;

pub use self::router::Router;

pub use self::log::Log;

pub use self::router::TreeRouter;

Modules

cache
Traits and implementations for cached resources.

context
Handler context and request body reading extensions.

handler
Request handlers.

header
Headers container, and common header fields.

log
Log tools.

mime

plugin
Request and context plugins.

response
Response writers.

router
Routers stores request handlers, using an HTTP method and a path as keys.

server
Server configuration and instance.

Macros

content_type!
A macro for assigning content types.

insert_routes!
The insert_routes! macro generates routes from the provided handlers and routing tree and adds them to the provided router. The router is then returned.

Enums

HttpError
A set of errors that can occur parsing HTTP streams.

HttpVersion
Represents a version of the HTTP spec.

Method
The Request Method (VERB)

Scheme
HTTP or HTTPS

StatusCode
An HTTP status code (status-code in RFC 7230 et al.).

Type Definitions

HttpResult