Struct rustful::cache::CachedProcessedFile

pub struct CachedProcessedFile<'p, T> {
    // some fields omitted
}

A processed cached file.

The file will be processed by a provided function each time it is loaded and the result will be stored.

use std::io;
use std::path::Path;
use std::fs::{File, PathExt};
use rustful::cache::{CachedValue, CachedProcessedFile};

fn get_size(_log: &Log, file: io::Result<File>) -> io::Result<Option<u64>> {
    file.and_then(|mut file| file.metadata()).map(|m| Some(m.len()))
}

let file = CachedProcessedFile::new(&Path::new("/some/file/path.txt"), None, get_size);
let guard = file.borrow(log);
match *guard {
    Some(ref size) => println!("file contains {} bytes of data", size),
    None => println!("the file was not loaded")
}

Methods

impl<'p, T: Send + Sync> CachedProcessedFile<'p, T>

fn new(path: &'p Path, unused_after: Option<u32>, processor: fn(&Log, Result<File>) -> Result<Option<T>>) -> CachedProcessedFile<'p, T>

Creates a new CachedProcessedFile which will be freed unused_after seconds after the latest access. The file will be processed by the provided processor function each time it's loaded.