Trait rustful::cache::CachedValue

pub trait CachedValue<'a, Value> {
    fn borrow_current(&'a self, log: &Log) -> Value;
    fn load(&self, log: &Log);
    fn free(&self, log: &Log);
    fn expired(&self, log: &Log) -> bool;
    fn unused(&self, log: &Log) -> bool;

    fn borrow(&'a self, log: &Log) -> Value { ... }
    fn clean(&self, log: &Log) { ... }
}

This trait provides functions for handling cached resources.

Required Methods

fn borrow_current(&'a self, log: &Log) -> Value

Borrow the cached value, without loading or reloading it.

fn load(&self, log: &Log)

Load the cached value.

fn free(&self, log: &Log)

Free the cached value.

fn expired(&self, log: &Log) -> bool

Check if the cached value has expired.

fn unused(&self, log: &Log) -> bool

Check if the cached value is unused and should be removed.

Provided Methods

fn borrow(&'a self, log: &Log) -> Value

Reload the cached value if it has expired and borrow it.

fn clean(&self, log: &Log)

Free the cached value if it's unused.