Trait rustful::context::body::ExtJsonBody

pub trait ExtJsonBody {
    fn read_json_body(&mut self) -> Result<Json, BuilderError>;
    fn decode_json_body<T: Decodable>(&mut self) -> DecodeResult<T>;
}

BodyReader extension for reading and parsing a JSON body.

It is available by default and can be toggled using the rustc_json_body feature.

Required Methods

fn read_json_body(&mut self) -> Result<Json, BuilderError>

Read the request body into a generic JSON structure. This structure can then be navigated and parsed freely.

A simplified example of how to parse { "a": number, "b": number }:

use rustful::{Context, Response};
use rustful::context::body::ExtJsonBody;

fn my_handler(mut context: Context, response: Response) {
    //Parse the request body as JSON
    let json = context.body.read_json_body().unwrap();

    //Find "a" and "b" in the root object and assume that they are numbers
    let a = json.find("a").and_then(|number| number.as_f64()).unwrap();
    let b = json.find("b").and_then(|number| number.as_f64()).unwrap();

    response.send(format!("{} + {} = {}", a, b, a + b));
}

fn decode_json_body<T: Decodable>(&mut self) -> DecodeResult<T>

Read and decode a request body as a type T. The target type must implement rustc_serialize::Decodable.

A simplified example of how to parse { "a": number, "b": number }:

extern crate rustful;
extern crate rustc_serialize;

use rustful::{Context, Response};
use rustful::context::body::ExtJsonBody;

#[derive(RustcDecodable)]
struct Foo {
    a: f64,
    b: f64
}

fn my_handler(mut context: Context, response: Response) {
    //Decode a JSON formatted request body into Foo
    let foo: Foo = context.body.decode_json_body().unwrap();

    response.send(format!("{} + {} = {}", foo.a, foo.b, foo.a + foo.b));
}