Trait rustful::context::body::ExtQueryBody

pub trait ExtQueryBody {
    fn read_query_body(&mut self) -> Result<Parameters>;
}

BodyReader extension for reading and parsing a query string.

Required Methods

fn read_query_body(&mut self) -> Result<Parameters>

Read and parse the request body as a query string. The body will be decoded as UTF-8 and plain '+' characters will be replaced with spaces.

A simplified example of how to parse a=number&b=number:

use rustful::{Context, Response};
use rustful::context::body::ExtQueryBody;

fn my_handler(mut context: Context, response: Response) {
    //Parse the request body as a query string
    let query = context.body.read_query_body().unwrap();

    //Find "a" and "b" and assume that they are numbers
    let a: f64 = query.get("a").and_then(|number| number.parse().ok()).unwrap();
    let b: f64 = query.get("b").and_then(|number| number.parse().ok()).unwrap();

    response.send(format!("{} + {} = {}", a, b, a + b));
}