Enum rustful::context::MaybeUtf8

pub enum MaybeUtf8<S, V> {
    Utf8(S),
    NotUtf8(V),
}

String data that may or may not be UTF-8 encoded.

Variants
Utf8(S)

A UTF-8 encoded string.

NotUtf8(V)

A non-UTF-8 string.

Methods

impl<S, V> MaybeUtf8<S, V>

pub fn new() -> MaybeUtf8<S, V> where
S: From<&'static str>,

Create an empty UTF-8 string.

pub fn as_slice<Sref: ?Sized, Vref: ?Sized>(&self) -> MaybeUtf8<&Sref, &Vref> where
S: AsRef<Sref>,
V: AsRef<Vref>,

Produce a slice of this string.

use rustful::context::{MaybeUtf8Owned, MaybeUtf8Slice};

let owned = MaybeUtf8Owned::from("abc");
let slice: MaybeUtf8Slice = owned.as_slice();

pub fn as_utf8(&self) -> Option<&str> where
S: AsRef<str>,

Borrow the string if it's encoded as valid UTF-8.

use rustful::context::MaybeUtf8Owned;

let string = MaybeUtf8Owned::from("abc");
assert_eq!(Some("abc"), string.as_utf8());

pub fn as_utf8_lossy(&self) -> Cow<str> where
S: AsRef<str>,
V: AsRef<[u8]>,

Borrow the string if it's encoded as valid UTF-8, or make a lossy conversion.

use rustful::context::MaybeUtf8Owned;

let string = MaybeUtf8Owned::from("abc");
assert_eq!("abc", string.as_utf8_lossy());

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn as_bytes(&self) -> &[u8] where
S: AsRef<[u8]>,
V: AsRef<[u8]>,

Borrow the string as a slice of bytes.

use rustful::context::MaybeUtf8Owned;

let string = MaybeUtf8Owned::from("abc");
assert_eq!(b"abc", string.as_bytes());

pub fn is_utf8(&self) -> bool

Check if the string is valid UTF-8.

use rustful::context::MaybeUtf8Owned;

let valid = MaybeUtf8Owned::from("abc");
assert_eq!(valid.is_utf8(), true);

let invalid = MaybeUtf8Owned::from(vec![255]);
assert_eq!(invalid.is_utf8(), false);

impl MaybeUtf8<String, Vec<u8>>

pub fn push_char(&mut self, c: char)

Push a single char to the end of the string.

use rustful::context::MaybeUtf8Owned;

let mut string = MaybeUtf8Owned::from("abc");
string.push_char('d');
assert_eq!("abcd", string);

pub fn push_byte(&mut self, byte: u8)

Push a single byte to the end of the string. The string's UTF-8 compatibility will be reevaluated and may change each time push_byte is called. This may have a noticeable performance impact.

use rustful::context::MaybeUtf8Owned;

let mut string = MaybeUtf8Owned::from("abc");
string.push_byte(255);
assert_eq!(string.is_utf8(), false);

pub fn push_str(&mut self, string: &str)

Extend the string.

use rustful::context::MaybeUtf8Owned;

let mut string = MaybeUtf8Owned::from("abc");
string.push_str("def");
assert_eq!("abcdef", string);

pub fn push_bytes(&mut self, bytes: &[u8])

Push a number of bytes to the string. The string's UTF-8 compatibility may change.

use rustful::context::MaybeUtf8Owned;

let mut string = MaybeUtf8Owned::from("abc");
string.push_bytes(&[100, 101, 102]);
assert_eq!("abcdef", string);

pub fn as_buffer(&mut self) -> Buffer

Borrow this string as a mutable byte buffer. The string's UTF-8 compatibility will be reevaluated when the buffer is dropped.