Struct rustful::context::hypermedia::Hypermedia

pub struct Hypermedia<'a> {
    pub links: Vec<Link<'a>>,
}

Hypermedia connected to an API endpoint.

Fields

links

Forward links from the current endpoint to other endpoints. These may include endpoints with the same path, but different method.

Methods

impl<'a> Hypermedia<'a>

fn new() -> Hypermedia<'a>

Create an empty Hypermedia structure.