Struct rustful::context::Buffer

pub struct Buffer<'a> {
    // some fields omitted
}

A byte buffer for more efficient MaybeUtf8 manipulation.

The buffer is essentially a &mut Vec<u8> that will be checked for UTF-8 compatibility when dropped. It comes with a few extra convenience methods.

Methods

impl<'a> Buffer<'a>

fn push_bytes(&mut self, bytes: &[u8])

Push a number of bytes to the buffer in a relatively efficient way.

use rustful::context::MaybeUtf8Owned;

let mut string = MaybeUtf8Owned::new();

{
    let mut buffer = string.as_buffer();
    buffer.push_bytes("abc".as_bytes());
}

assert!(string.is_utf8());
assert_eq!("abc", string);

fn push_char(&mut self, c: char)

Push a single char to the end of the buffer.

use rustful::context::MaybeUtf8Owned;

let mut string = MaybeUtf8Owned::new();

{
    let mut buffer = string.as_buffer();
    buffer.push_char('å');
    buffer.push_char('1');
    buffer.push_char('€');
}

assert!(string.is_utf8());
assert_eq!("å1€", string);