Trait rustful::context::ExtJsonBody

pub trait ExtJsonBody {
    fn read_json_body(&mut self) -> Result<Json, BuilderError>;
    fn decode_json_body<T: Decodable>(&mut self) -> DecodeResult<T>;
}

BodyReader extension for reading and parsing a JSON body.

It is available by default and can be toggled using the rustc_json_body feature. Examples and more information can be found in the documentation for BodyReader.

Required Methods

fn read_json_body(&mut self) -> Result<Json, BuilderError>

Read the request body into a JSON structure.

fn decode_json_body<T: Decodable>(&mut self) -> DecodeResult<T>

Parse and decode the request body as some type T.