Enum rustful::Global

pub enum Global {
    // some variants omitted
}

Container for globally accessible data.

It should only be created using Global::default()/Default::default() or Box::new(v).into().

let g: Global = Box::new(5).into();
assert_eq!(g.get(), Some(&5));

Direct manipulation of its variants may cause an undesired state.

Note: This structure can currently only hold either one item or nothing. A future version will be able to hold multiple items.

Methods

impl Global

fn get<T: Any + Send + Sync>(&self) -> Option<&T>

Borrow a value of type T if the there is one.