Function rustful::file::ext_to_mime

pub fn ext_to_mime(ext: &str) -> Option<Mime>

Returns the MIME type from a given file extension, if known.

The file extension to MIME type mapping is based on data from the Apache server.

use rustful::file::ext_to_mime;
use rustful::mime::Mime;
use rustful::mime::TopLevel::Image;
use rustful::mime::SubLevel::Jpeg;

let mime = ext_to_mime("jpg");
assert_eq!(mime, Some(Mime(Image, Jpeg, vec![])));