Struct rustful::file::Loader

pub struct Loader {
    pub buffer_size: usize,
}

A utility for loading files from the file system and sending them to the client.

Fields

buffer_size

The size, in bytes, of the file chunks. Default is 1048576 (1 megabyte).

Methods

impl Loader

fn new() -> Loader

fn send_file<'a, 'b, P: AsRef<Path>>(&self, path: P, response: Response<'a, 'b>) -> Result<(), Error<'a, 'b>>

Send a file to the client.

A MIME type is automatically applied to the response, based on the file extension, and application/octet-stream is used as a fallback if the extension is unknown. See ext_to_mime for more information.