Struct rustful::handler::or_else::OrElse

pub struct OrElse<A, B> {
    pub primary: A,
    pub secondary: B,
}

A router that selects a secondary handler on error.

use rustful::{OrElse, Context, Response};

fn error_handler(_: Context, response: Response) {
    let status = response.status();
    response.send(format!("Status: {}", status));
}

//Every request will end up at error_handler
let handler = OrElse::<Option<fn(Context, Response)>, _>::new(None, error_handler);

Fields
primary: A

The primary handler that will have the first chance to process the request.

secondary: B

The secondary handler that will take over if the first handler returns an error.

Methods

impl<A, B> OrElse<A, B>

pub fn new(primary: A, secondary: B) -> OrElse<A, B>

Create a new OrElse with a primary and secondary handler.

pub fn build(&mut self) -> Builder<A, B>

Build the router and its children using a chaninable API.

use rustful::{Context, Response, OrElse};
use rustful::handler::MethodRouter;

type Inner = MethodRouter<fn(Context, Response)>;

fn get(_context: Context, response: Response) {
    response.send("A GET request.");
}

fn post(_context: Context, response: Response) {
    response.send("A POST request.");
}

fn or_else(_context: Context, response: Response) {
    response.send("Hello world!");
}

let mut router = OrElse::<Inner, _>::with_secondary(or_else as fn(Context, Response));

router.build().primary().many(|mut inner_router|{
    inner_router.on_get(get as fn(Context, Response));
    inner_router.on_post(post);
});

impl<A, B: Default> OrElse<A, B>

pub fn with_primary(primary: A) -> OrElse<A, B>

Create a new OrElse with a given primary and a default secondary handler.

impl<A: Default, B> OrElse<A, B>

pub fn with_secondary(secondary: B) -> OrElse<A, B>

Create a new OrElse with a given secondary and a default primary handler.