Struct rustful::handler::routing::RouteState

pub struct RouteState<'a> { /* fields omitted */ }

A state object for routing.

Methods

impl<'a> RouteState<'a>

fn get(&self) -> Option<&'a [u8]>

Get the current path segment.

fn skip(&mut self)

Don't include this path segment in a variable.

fn keep(&mut self)

Include this path segment as a variable.

fn fuse(&mut self)

Extend a previously saved variable value with this path segment, or save it as a new variable.

fn variables(&self,
names: &[MaybeUtf8Owned])
-> HashMap<MaybeUtf8Owned, MaybeUtf8Owned>

Assign names to the saved variables and return them.

fn snapshot(&self) -> (usize, usize)

Get a snapshot of a part of the current state.

fn go_to(&mut self, snapshot: (usize, usize))

Go to a previously recorded state.

fn is_empty(&self) -> bool

Check if there are no more segments.