Trait rustful::handler::routing::Insert

pub trait Insert<H> {
    fn build<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(method: Method,
                                                                          route: R,
                                                                          item: H)
                                                                          -> Self;
    fn insert<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(&mut self,
                                                                           method: Method,
                                                                           route: R,
                                                                           item: H);
}

A common trait for building routers.

A router has to implement this trait to be compatible with the insert_routes! macro.

Required Methods

fn build<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(method: Method,
route: R,
item: H)
-> Self

Build a new router from a route. The router may choose to ignore both method and route, depending on its implementation.

fn insert<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(&mut self,
method: Method,
route: R,
item: H)

Insert a new route into the router. The router may choose to ignore both method and route, depending on its implementation.