Struct rustful::handler::status_router::StatusRouter

pub struct StatusRouter<T> { /* fields omitted */ }

A router that selects an item from an HTTP status code.

It does neither alter nor preserve the status code of the response.

Methods

impl<T> StatusRouter<T>

fn new() -> StatusRouter<T>

Create an empty StatusRouter.

fn build(&mut self) -> Builder<T>

Build the router and its children using a chaninable API.

use rustful::{Context, Response, StatusCode, StatusRouter};

fn on_404(context: Context, response: Response) {
    response.send(format!("{} was not found.", context.uri_path));
}

fn on_500(_: Context, response: Response) {
    response.send("Looks like you found a bug");
}
let mut status_router = StatusRouter::<fn(Context, Response)>::new();

status_router.build().many(|mut status_router|{
    status_router.on(StatusCode::NotFound, on_404 as fn(Context, Response));
    status_router.on(StatusCode::InternalServerError, on_500);
});

fn insert(&mut self, status: StatusCode, handler: T)

Insert a handler that will listen for a specific status code.

use rustful::{Context, Response, StatusCode, StatusRouter};
use rustful::handler::MethodRouter;

let on_not_found = MethodRouter::<fn(Context, Response)>::new();
//Fill on_not_found with handlers...

let mut status_router = StatusRouter::new();
status_router.insert(StatusCode::NotFound, on_not_found);