Struct rustful::handler::Environment

pub struct Environment<'a, 'b: 'a, 'l, 'g> {
    pub context: Context<'a, 'b, 'l, 'g>,
    pub response: Response<'a, 'g>,
    pub route_state: RouteState<'g>,
}

A request environment, containing the context, response and route state.

Fields
context: Context<'a, 'b, 'l, 'g>

The request context, containing the request parameters.

response: Response<'a, 'g>

The response that will be sent to the client.

route_state: RouteState<'g>

A representation of the current routing state.

Methods

impl<'a, 'b, 'l, 'g> Environment<'a, 'b, 'l, 'g>

fn replace_hyperlinks<'n>(self,
hyperlinks: Vec<Link<'n>>)
-> (Environment<'a, 'b, 'n, 'g>, Vec<Link<'l>>)

Replace the hyperlinks. This consumes the whole request environment and returns a new one with a different lifetime, together with the old hyperlinks.