Struct rustful::handler::Variables

pub struct Variables<H> { /* fields omitted */ }

Assigns names to route variables.

It's technically a single-item router, with the purpose of pairing route variable names with input values, but it has to come after other routers, since it has be at the end of a router chain to know what variables to use. It won't do any other routing work, so make sure to pair it with, at least, a path based router.