Trait rustful::handler::CreateContent

pub trait CreateContent: Send + Sync + 'static {
    type Output: for<'a, 'b> SendResponse<'a, 'b>;
    fn create_content(&self, context: Context) -> Self::Output;
}

A simplified handler that returns the response content.

It's meant for situations where direct access to the Response is unnecessary. The benefit is that it can use early returns in a more convenient way.

Associated Types

type Output: for<'a, 'b> SendResponse<'a, 'b>

The type of content that is created.

Required Methods

fn create_content(&self, context: Context) -> Self::Output

Create content that will be sent to the client.