Trait rustful::handler::HandleRequest

pub trait HandleRequest: Send + Sync + 'static {
    fn handle_request<'a, 'b, 'l, 'g>(
        &self, 
        environment: Environment<'a, 'b, 'l, 'g>
    ) -> Result<(), Environment<'a, 'b, 'l, 'g>>;
    fn hyperlinks<'a>(&'a self, base_link: Link<'a>) -> Vec<Link<'a>>;
}

A low level request handler.

Suitable for implementing routers, since it has complete access to the routing environment and can reject the request.

Required Methods

fn handle_request<'a, 'b, 'l, 'g>(
&self,
environment: Environment<'a, 'b, 'l, 'g>
) -> Result<(), Environment<'a, 'b, 'l, 'g>>

Try to handle an incoming request from the client, or return the request environment to the parent handler.

fn hyperlinks<'a>(&'a self, base_link: Link<'a>) -> Vec<Link<'a>>

List all of the hyperlinks into this handler, based on the provided base link. It's up to the handler implementation to decide how deep to go.