Struct rustful::handler::tree_router::TreeRouter

pub struct TreeRouter<T> {
    pub find_hyperlinks: bool,
    // some fields omitted
}

A tree shaped router that selects handlers using paths.

Each tree node stores an other router of type T, which has to implement Default to allow "empty" nodes, and a number of child references. The requested path must always be exhausted before sub-routers are searched, so there is no point in storing other path based routers in a TreeRouter.

The TreeRouter has support for shallow hyperlinks to children, siblings, cousins, and so forth. The use of variable sequences complicates this process and may cause confusing results in certain situations. The hyperlinks may or may not point to a handler.

Hyperlinks has to be activated by setting find_hyperlinks to true.

Fields
find_hyperlinks: bool

Should the router search for hyperlinks? Setting this to true may slow down endpoint search, but enables hyperlinks.

Methods

impl<T: Default> TreeRouter<T>

fn new() -> TreeRouter<T>

Creates an empty TreeRouter.

impl<T> TreeRouter<T>

fn with_handler(handler: T) -> TreeRouter<T>

Creates a TreeRouter with only a root handler.

use rustful::{Context, Response};
use rustful::handler::TreeRouter;

fn handler(_context: Context, response: Response) {
    response.send("Hello world!");
}

let router = TreeRouter::with_handler(handler);

fn build(&mut self) -> Builder<T>

Build the router and its children using a chaninable API.

use rustful::{Context, Response};
use rustful::handler::TreeRouter;

fn handler(_context: Context, response: Response) {
    response.send("Hello world!");
}

let mut router = TreeRouter::<Option<fn(Context, Response)>>::new();
router.build().on_path("hello/world", handler);