Module rustful::headerExperimental

Headers container, and common header fields.

hyper has the opinion that Headers should be strongly-typed, because that's why we're using Rust in the first place. To set or get any header, an object must implement the Header trait from this module. Several common headers are already provided, such as Host, ContentType, UserAgent, and others.

Modules

accept
Exposes the Accept header.

allow
Exposes the Allow header.

authorization
Exposes the Authorization header.

cache_control
Exposes the CacheControl header.

common
Common Headers A Collection of Header implementations for common HTTP Headers.

connection
Exposes the Connection header.

content_length
Exposes the ContentLength header.

content_type
Exposes the ContentType header.

cookie
Exposes the Cookie header.

date
Exposes the Date header.

etag
Exposes the Etag header.

expires
Exposes the Expires header.

host
Exposes the Host header.

if_modified_since
Exposes the If-Modified-Since header.

last_modified
Exposes the LastModified header.

location
Exposes the Location header.

server
Exposes the Server header.

set_cookie
Exposes the Set-Cookie header.

transfer_encoding
Exposes the TransferEncoding header.

upgrade
Exposes the Upgrade header.

user_agent
Exposes the UserAgent header.

util
Utility functions for Header implementations.

vary
Exposes the Vary header.

Structs

Accept
The Accept header.

Allow
The Allow header. See also https://tools.ietf.org/html/rfc7231#section-7.4.1

Authorization
The Authorization header field.

CacheControl
The Cache-Control header.

CaseInsensitive
Case-insensitive string.

Connection
The Connection header.

ContentLength
The Content-Length header.

ContentType
The Content-Type header.

Cookies
The Cookie header. Defined in RFC6265:

Date
The Date header field.

Etag
The Etag header.

Expires
The Expires header field.

HeaderFormatter
A wrapper around any Header with a Show impl that calls fmt_header.

HeaderView
Returned with the HeadersItems iterator.

Headers
A map of header fields on requests and responses.

HeadersItems
An Iterator over the fields in a Headers map.

Host
The Host header.

IfModifiedSince
The If-Modified-Since header field.

LastModified
The LastModified header field.

Location
The Location header.

Server
The Server header field.

SetCookie
The Set-Cookie header

TransferEncoding
The Transfer-Encoding header.

Upgrade
The Upgrade header.

UserAgent
The User-Agent header field.

Enums

Vary
The Allow header. See also https://tools.ietf.org/html/rfc7231#section-7.1.4

Traits

Header
A trait for any object that will represent a header field and value.

HeaderFormat
A trait for any object that will represent a header field and value.