Module rustful::header::commonExperimental

Common Headers A Collection of Header implementations for common HTTP Headers.

Mime

Several header fields use MIME values for their contents. Keeping with the strongly-typed theme, the mime crate is used, such as ContentType(pub Mime).

Modules

accept
Exposes the Accept header.

allow
Exposes the Allow header.

authorization
Exposes the Authorization header.

cache_control
Exposes the CacheControl header.

connection
Exposes the Connection header.

content_length
Exposes the ContentLength header.

content_type
Exposes the ContentType header.

cookie
Exposes the Cookie header.

date
Exposes the Date header.

etag
Exposes the Etag header.

expires
Exposes the Expires header.

host
Exposes the Host header.

if_modified_since
Exposes the If-Modified-Since header.

last_modified
Exposes the LastModified header.

location
Exposes the Location header.

server
Exposes the Server header.

set_cookie
Exposes the Set-Cookie header.

transfer_encoding
Exposes the TransferEncoding header.

upgrade
Exposes the Upgrade header.

user_agent
Exposes the UserAgent header.

util
Utility functions for Header implementations.

Structs

Accept
The Accept header.

Allow
The Allow header. See also https://tools.ietf.org/html/rfc7231#section-7.4.1

Authorization
The Authorization header field.

CacheControl
The Cache-Control header.

Connection
The Connection header.

ContentLength
The Content-Length header.

ContentType
The Content-Type header.

Cookies
The Cookie header. Defined in RFC6265:

Date
The Date header field.

Etag
The Etag header.

Expires
The Expires header field.

Host
The Host header.

IfModifiedSince
The If-Modified-Since header field.

LastModified
The LastModified header field.

Location
The Location header.

Server
The Server header field.

SetCookie
The Set-Cookie header

TransferEncoding
The Transfer-Encoding header.

Upgrade
The Upgrade header.

UserAgent
The User-Agent header field.