Trait rustful::log::Log

pub trait Log: Send + Sync {
    fn try_note(&self, message: &str) -> Result;
    fn try_warning(&self, message: &str) -> Result;
    fn try_error(&self, message: &str) -> Result;

    fn note(&self, message: &str) { ... }
    fn warning(&self, message: &str) { ... }
    fn error(&self, message: &str) { ... }
}

Common trait for log tools.

Required Methods

fn try_note(&self, message: &str) -> Result

Print a note to the log or return eventual errors.

fn try_warning(&self, message: &str) -> Result

Print a warning to the log or return eventual errors.

fn try_error(&self, message: &str) -> Result

Print an error to the log or return eventual errors.

Provided Methods

fn note(&self, message: &str)

Print a note to the log and ignore any errors.

fn warning(&self, message: &str)

Print a warning to the log and ignore any errors.

fn error(&self, message: &str)

Print an error to the log and ignore any errors.