Struct rustful::plugin::PluginContext

pub struct PluginContext<'a> {
    pub storage: &'a mut AnyMap,
    pub log: &'a Log,
}

Contextual tools for plugins.

Fields

storage

Shared storage for plugins. It is local to the current request and accessible from the handler and all of the plugins. It can be used to send data between these units.

log

Log for notes, errors and warnings.