Enum rustful::response::FileError

pub enum FileError<'a, 'b> {
    Open(Error, Response<'a, 'b>),
    Send(Error),
}

Error that may occure while sending a file.

Variants

Open

Failed to open the file.

Send

Failed while sending the file.

Methods

impl<'a, 'b> FileError<'a, 'b>

fn recover_response(self) -> Result<Response<'a, 'b>, FileError<'a, 'b>>

Recover the response if the file couldn't be opened.

fn send_not_found<'d, M: Into<Data<'d>>>(self, message: M) -> Result<(), FileError<'a, 'b>>

Send a 404 (not found) response if the file wasn't found, or return self if any other error occurred.

fn ignore_send_error(self) -> Result<(), (Error, Response<'a, 'b>)>

Ignore any error that might have occurred while sending the file.