Enum rustful::response::ResponseData

pub enum ResponseData<'a> {
    Bytes(Vec<u8>),
    ByteSlice(&'a [u8]),
    String(String),
    StringSlice(&'a str),
}

Variants

Bytes

Data in byte form.

ByteSlice

Data in byte form.

String

Data in string form.

StringSlice

Data in string form.

Methods

impl<'a> ResponseData<'a>

fn as_bytes(&self) -> &[u8]

Borrow the content as a byte slice.

fn into_bytes(self) -> Vec<u8>

Turns the content into a byte vector. Slices are copied.

fn as_string(&self) -> Option<&str>

Borrow the content as a string slice if the content is a string. Returns an None if the content is a byte vector, a byte slice or if the action is Error.

fn into_string(self) -> Option<String>

Extract the contained string or string slice if there is any. Returns an None if the content is a byte vector, a byte slice or if the action is Error. Slices are copied.