Struct rustful::response::Chunked

pub struct Chunked<'a, 'b> {
    // some fields omitted
}

An interface for writing a chunked response body.

This is useful for when the size of the data is unknown, but it comes with an overhead for each time send or try_send is called (simply put).

Methods

impl<'a, 'b> Chunked<'a, 'b>

fn filter_storage(&self) -> &AnyMap

Get a reference to the filter storage.

fn filter_storage_mut(&mut self) -> &mut AnyMap

Get a mutable reference to the filter storage. It can be used to communicate with the response filters.

fn send<'d, Content: Into<Data<'d>>>(&mut self, content: Content)

Send a chunk of data to the client, ignoring any eventual errors. Use try_send to get error information.

use rustful::{Context, Response};

fn my_handler(context: Context, response: Response) {
    let count = context.variables.get("count")
        .and_then(|n| n.parse().ok())
        .unwrap_or(0u32);
    let mut chunked = response.into_chunked();

    for i in 0..count {
        chunked.send(format!("chunk #{}", i + 1));
    }
}

fn try_send<'d, Content: Into<Data<'d>>>(&mut self, content: Content) -> Result<usize, Error>

Send a chunk of data to the client. This is the same as send, but errors are not ignored.

#[macro_use] extern crate log;
use rustful::{Context, Response};
use rustful::response::Error;

fn my_handler(context: Context, response: Response) {
    let count = context.variables.get("count")
        .and_then(|n| n.parse().ok())
        .unwrap_or(0u32);
    let mut chunked = response.into_chunked();

    for i in 0..count {
        if let Err(Error::Filter(e)) = chunked.try_send(format!("chunk #{}", i + 1)) {
            error!("a filter failed: {}", e);
        }
    }
}

fn end(self) -> Result<(), Error>

Finish writing the response and collect eventual errors.

This is optional and will happen silently when the writer drops out of scope.