Struct rustful::response::ResponseWriter

pub struct ResponseWriter<'a, 'b> {
    // some fields omitted
}

An interface for writing to the response body.

Methods

impl<'a, 'b> ResponseWriter<'a, 'b>

fn plugin_storage(&mut self) -> &mut AnyMap

Mutably borrow the plugin storage. It can be used to communicate with the response plugins.

fn send<'d, Content: Into<Data<'d>>>(&mut self, content: Content) -> Result<usize, Error>

Writes response body data to the client.

fn end(self) -> Result<(), Error>

Finish writing the response and collect eventual errors.

This is optional and will happen when the writer drops out of scope.