Module rustful::router

Routers stores request handlers, using an HTTP method and a path as keys.

The TreeRouter, for example, can be created from a vector of predefined paths, like this:

let routes = vec![
    (Get, "/about", about_us as fn(Context, Response)),
    (Get, "/user/:user", show_user as fn(Context, Response)),
    (Get, "/product/:name", show_product as fn(Context, Response)),
    (Get, "/*", show_error as fn(Context, Response)),
    (Get, "/", show_welcome as fn(Context, Response))
];

let router = TreeRouter::from_routes(routes);

Routes may also be added after the router was created, like this:

let mut router = TreeRouter::new();

router.insert(Get, "/about", about_us as fn(Context, Response));
router.insert(Get, "/user/:user", show_user as fn(Context, Response));
router.insert(Get, "/product/:name", show_product as fn(Context, Response));
router.insert(Get, "/*", show_error as fn(Context, Response));
router.insert(Get, "/", show_welcome as fn(Context, Response));

Routers can also be used with a special macro for creating routes, called insert_routes!:

#![feature(phase)]
#[phase(plugin)]
extern crate rustful_macros;

extern crate rustful;

...

let router = insert_routes!{
    TreeRouter::new(): {
        "/about" => Get: about_us as fn(Context, Response),
        "/user/:user" => Get: show_user as fn(Context, Response),
        "/product/:name" => Get: show_product as fn(Context, Response),
        "/*" => Get: show_error as fn(Context, Response),
        "/" => Get: show_welcome as fn(Context, Response)
    }
};

This macro will generate the same code as the example above, but it allows more advanced structures to be defined without the need to write the same paths multiple times. This can be useful to lower the risk of typing errors, among other things.

This can be found in the crate rustful_macros.

Structs

TreeRouter
Stores items, such as request handlers, using an HTTP method and a path as keys.

Traits

Router
Routers are used to store request handlers.

Type Definitions

RouterResult