Struct rustful::router::MethodRouter

pub struct MethodRouter<T> { /* fields omitted */ }

A router that selects an item from an HTTP method.

It's a simple mapping between Method and a router T, while the requested path is ignored. It's therefore a good idea to pair a MethodRouter with an exhaustive path router of some sort.