Trait rustful::router::Router

pub trait Router: Send + Sync + 'static {
    type Handler: Handler;
    fn build<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(method: Method, route: R, item: Self::Handler) -> Self;
    fn insert<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(&mut self, method: Method, route: R, item: Self::Handler);
    fn insert_router<'a, R: Into<InsertState<'a, I>>, I: Clone + Iterator<Item=&'a [u8]>>(&mut self, route: R, router: Self);
    fn prefix<'a, R: Into<InsertState<'a, I>>, I: Clone + Iterator<Item=&'a [u8]>>(&mut self, route: R);
    fn find<'a>(&'a self, method: &Method, route: &mut RouteState) -> Endpoint<'a, Self::Handler>;
    fn hyperlinks<'a>(&'a self, base: Link<'a>) -> Vec<Link<'a>>;

    fn merge(&mut self, other: Self) where Self: Sized { ... }
}

A common trait for routers.

A router must to implement this trait to be usable in a Rustful server. This trait will also make the router compatible with the insert_routes! macro.

Associated Types

type Handler: Handler

The request handler type that is stored within this router.

Required Methods

fn build<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(method: Method, route: R, item: Self::Handler) -> Self

Build a new router from a route. The router may choose to ignore both method and route, depending on its implementation.

fn insert<'a, R: Into<InsertState<'a, I>>, I: Iterator<Item=&'a [u8]>>(&mut self, method: Method, route: R, item: Self::Handler)

Insert a new route into the router. The router may choose to ignore both method and route, depending on its implementation.

fn insert_router<'a, R: Into<InsertState<'a, I>>, I: Clone + Iterator<Item=&'a [u8]>>(&mut self, route: R, router: Self)

Insert an other router at a path. The content of the other router will be merged with this one and conflicting content will be overwritten.

fn prefix<'a, R: Into<InsertState<'a, I>>, I: Clone + Iterator<Item=&'a [u8]>>(&mut self, route: R)

Change the router as if it was placed under the provided route.

fn find<'a>(&'a self, method: &Method, route: &mut RouteState) -> Endpoint<'a, Self::Handler>

Find and return the matching handler and variable values.

fn hyperlinks<'a>(&'a self, base: Link<'a>) -> Vec<Link<'a>>

List all of the hyperlinks into this router, based on the provided base link. It's up to the router implementation to decide how deep to go.

Provided Methods

fn merge(&mut self, other: Self) where Self: Sized

Merge this router with an other one, overwriting conflicting parts.