Struct rustful::server::Global

pub struct Global(_);

A somewhat lazy container for globally accessible data.

It will try to be as simple as possible and allocate as little as possible, depending on the number of stored values.

No value: Nothing is allocated and nothing is searched for during access.

One value: One Box is allocated. Searching for a value will only consist of a comparison of TypeId and a downcast.

Multiple values: An AnyMap is created, as well as a Box for each value. Searching for a value has the full overhead of AnyMap.

Global can be created from a boxed value, from tuples or using the Default trait. More values can then be added using insert(value).

use rustful::server::Global;
let mut g1: Global = Box::new(5).into();
assert_eq!(g1.get(), Some(&5));
assert_eq!(g1.get::<&str>(), None);

let old = g1.insert(10);
assert_eq!(old, Some(5));
assert_eq!(g1.get(), Some(&10));

g1.insert("cat");
assert_eq!(g1.get(), Some(&10));
assert_eq!(g1.get(), Some(&"cat"));

let g2: Global = (5, "cat").into();
assert_eq!(g2.get(), Some(&5));
assert_eq!(g2.get(), Some(&"cat"));

Methods

impl Global

pub fn get<T: Any + Send + Sync>(&self) -> Option<&T>

Borrow a value of type T if the there is one.

pub fn insert<T: Any + Send + Sync>(&mut self, value: T) -> Option<T>

Insert a new value, returning the previous value of the same type, if any.