Struct rustful::server::ServerInstance

pub struct ServerInstance<R: Router> {
    // some fields omitted
}

A runnable instance of a server.

It's not meant to be used directly, unless additional control is required.

let (server_instance, scheme) = Server {
    host: 8080.into(),
    handlers: router,
    ..Server::default()
}.build();

Methods

impl<R: Router> ServerInstance<R>

fn new(config: Server<R>) -> (ServerInstance<R>, Scheme)

Create a new server instance, with the provided configuration. This is the same as Server{...}.build().

fn run(self, _scheme: Scheme) -> HttpResult<Listening>

Start the server.