Struct rustful::ListeningExperimental

pub struct Listening<A = HttpAcceptor> {
    pub socket: SocketAddr,
    // some fields omitted
}

A listening server, which can later be closed.

Fields

socket

The socket addresses that the server is bound to.

Methods

impl<A: NetworkAcceptor<S>, S: NetworkStream> Listening<A>

fn close(&mut self) -> Result<(), HttpError>

Stop the server from listening to its socket address.