Struct rustful::Response

pub struct Response<'a, 'b, 'c> {
    pub headers: HeaderCollection,
    pub status: Status,
    // some fields omitted
}

An interface for sending HTTP response data to the client.

Fields

headers

The HTTP response headers. Date, content type (text/plain) and server is automatically set.

status

The HTTP response status. Ok (200) is default.

Methods

impl<'a, 'b, 'c> Response<'a, 'b, 'c>

fn send<'a, Content: IntoResponseData<'a>>(&mut self, content: Content) -> ResponseResult

Writes a string or any other BytesContainer to the client. The headers will be written the first time send() is called.

fn begin(&mut self) -> ResponseResult

Start writing the response. Headers and status can not be changed after it has been called.

This method will be called automatically by send() and end(), if it hasn't been called before. It can only be called once.

fn end(&mut self) -> ResponseResult

Finish writing the response.