Struct rustful::Server

pub struct Server<R, C> {
    pub handlers: R,
    pub port: Port,
    pub host: IpAddr,
    pub protocol: Protocol,
    pub threads: Option<usize>,
    pub server: String,
    pub content_type: Mime,
    pub cache: C,
    pub cache_clean_interval: Option<i64>,
    pub context_plugins: Vec<Box<ContextPlugin<Cache=C> + Send + Sync>>,
    pub response_plugins: Vec<Box<ResponsePlugin + Send + Sync>>,
}

Used to build and run a server.

Each field has a corresponding modifier method and calls to these methods can be chained for quick setup.

let server_result = Server::new().port(8080).handlers(router).run();

match server_result {
    Ok(_server) => {},
    Err(e) => println!("could not start server: {}", e.description())
}

Fields

handlers

One or several response handlers.

port

The port where the server will listen for requests.

host

The host address where the server will listen for requests.

protocol

The type of HTTP protocol.

threads

The number of threads to be used in the server task pool. The default (None) is based on recommendations from the system.

server

The content of the server header.

content_type

The default media type.

cache

A structure where resources are cached.

cache_clean_interval

How often the cache should be cleaned. Measured in seconds.

context_plugins

The context plugin stack.

response_plugins

The response plugin stack.

Methods

impl Server<(), ()>

fn new() -> Server<(), ()>

Create a new empty server which will listen on host address 0.0.0.0 and port 80.

impl<C: Cache> Server<(), C>

fn with_cache(cache: C) -> Server<(), C>

Create a new empty server with a cache.

impl<R, C> Server<R, C>

fn handlers<NewRouter: Router<Handler=H>, H: Handler<Cache=C>>(self, handlers: NewRouter) -> Server<NewRouter, C>

Set request handlers.

fn port(self, port: Port) -> Server<R, C>

Change the port. Default is 80.

fn host(self, host: IpAddr) -> Server<R, C>

Change the host address. Default is 0.0.0.0.

fn https(self, cert: Path, key: Path) -> Server<R, C>

Change the protocol to HTTPS.

fn threads(self, threads: Option<usize>) -> Server<R, C>

Set the number of threads to be used in the server task pool.

Passing None will set it to the default number of threads, based on recommendations from the system.

fn cache_clean_interval(self, interval: Option<u32>) -> Server<R, C>

Set the minimal number of seconds between each cache clean.

Passing None disables cache cleaning.

fn server_name(self, name: String) -> Server<R, C>

Change the server response header. Default is rustful.

fn content_type(self, content_type: Mime) -> Server<R, C>

Change the default content type. Default is text/plain.

fn with_context_plugin<P: ContextPlugin<Cache=C> + Send + Sync>(self, plugin: P) -> Server<R, C>

Add a context plugin to the plugin stack.

fn with_response_plugin<P: ResponsePlugin + Send + Sync>(self, plugin: P) -> Server<R, C>

Add a response plugin to the plugin stack.

impl<R, H, C> Server<R, C> where R: Router<Handler=H> + Send + Sync, H: Handler<Cache=C> + Send + Sync, C: Cache + Send + Sync

fn run(self) -> HttpResult<Listening>

Start the server.

fn build(self) -> (ServerInstance<R, C>, Protocol)

Build a runnable instance of the server.