Struct rustful::StatusRouter

pub struct StatusRouter<T> { /* fields omitted */ }

A router that selects an item from an HTTP status code.

It does neither alter nor preserve the status code of the response.

use rustful::{OrElse, Context, Response, StatusCode, StatusRouter};

fn on_404(context: Context, response: Response) {
    response.send(format!("{} was not found.", context.uri_path));
}

fn on_500(_: Context, response: Response) {
    response.send("Looks like you found a bug");
}

fn on_unexpected(_: Context, response: Response) {
    let status = response.status();
    response.send(format!("Unexpected status: {}", status));
}

//Use error_handler.primary to catch specific errors and error_handler.secondary to
//catch everything else
let mut error_handler = OrElse::new(StatusRouter::new(), on_unexpected);
error_handler.primary.on_status(StatusCode::NotFound, on_404 as fn(Context, Response));
error_handler.primary.on_status(StatusCode::InternalServerError, on_500);

//Every request will end up at error_handler
let handler = OrElse::<Option<fn(Context, Response)>, _>::new(None, error_handler);

Methods

impl<T> StatusRouter<T>

fn new() -> StatusRouter<T>

Create an empty StatusRouter.

fn on_status(&mut self, status: StatusCode, handler: T)

Insert a handler that will listen for a specific status code.