Crate rustful

Reexports

pub use router::TreeRouter;

Modules

cache
Utility traits and implementations for cached resources.

request_extensions

router
TreeRouter stores items, such as request handlers, using an HTTP method and a path as keys.

Structs

Request
A container for all the request data, including get, set and path variables.

Response
An interface for sending HTTP response data to the client.

Server
Used to build and run a server.

ServerInstance
A runnable instance of a server.

Enums

RequestAction
The result from a RequestPlugin.

ResponseAction
The result from a ResponsePlugin.

ResponseData

ResponseError
The result of a response action.

Traits

Cache
A trait for cache storage.

Handler
A trait for request handlers.

IntoResponseData

RequestPlugin
A trait for request plugins.

ResponsePlugin
A trait for response plugins.

Router
Routers are used to store request handlers.