Struct rustful::cache::CachedProcessedFile

pub struct CachedProcessedFile<T> {
    // some fields omitted
}

A processed cached file.

The file will be processed by a provided function each time it is loaded and the result will be stored.

use std::io::{File, IoResult};
use rustful::cache::{CachedValue, CachedProcessedFile};

fn get_size(file: IoResult<File>) -> IoResult<Option<u64>> {
    file.and_then(|mut file| file.stat()).map(|stat| Some(stat.size))
}

let file = CachedProcessedFile::new(Path::new("/some/file/path.txt"), None, get_size);

match *file.borrow() {
    Some(ref size) => println!("file contains {} bytes of data", size),
    None => println!("the file was not loaded")
}

Methods

impl<T: Send + Sync> CachedProcessedFile<T>

fn new(path: Path, unused_after: Option<u32>, processor: fn(IoResult<File>) -> IoResult<Option<T>>) -> CachedProcessedFile<T>

Creates a new CachedProcessedFile which will be freed unused_after seconds after the latest access. The file will be processed by the provided processor function each time it's loaded.