Struct rustful::context::Context

pub struct Context<'a, 'b, 'l> {
    pub headers: Headers,
    pub http_version: HttpVersion,
    pub address: SocketAddr,
    pub method: Method,
    pub path: String,
    pub variables: HashMap<String, String>,
    pub query: HashMap<String, String>,
    pub fragment: Option<String>,
    pub log: &'l Log + 'l,
    pub body_reader: BodyReader<'a, 'b>,
}

A container for things like request data and cache.

A Context can be dereferenced to a BodyReader, allowing direct access to the underlying read methods.

Fields

headers

Headers from the HTTP request.

http_version

The HTTP version used in the request.

address

The client address

method

The HTTP method.

path

The requested path.

variables

Route variables.

query

Query variables from the path.

fragment

The fragment part of the URL (after #), if provided.

log

Log for notes, errors and warnings.

body_reader

A reader for the request body.