Struct rustful::context::Parameters

pub struct Parameters(_);

An extended HashMap with extra functionality for value parsing.

Some of the methods from HashMap has been wrapped to provide a more ergonomic API, where anything that can be represented as a byte slice can be used as a key.

Methods

impl Parameters

pub fn new() -> Parameters

Create an empty Parameters.

pub fn get<'a, K: ?Sized>(&'a self, key: &K) -> Option<Cow<'a, str>> where
K: Hash + Eq + AsRef<[u8]>,

Get a parameter as a UTF-8 string. A lossy conversion will be performed if it's not encoded as UTF-8. Use get_raw to get the original data.

pub fn get_raw<'a, K: ?Sized>(&'a self, key: &K) -> Option<&'a MaybeUtf8Owned> where
K: Hash + Eq + AsRef<[u8]>,

Get a parameter that may or may not be a UTF-8 string.

pub fn get_mut<'a, K: ?Sized>(
&'a mut self,
key: &K
) -> Option<&'a mut MaybeUtf8Owned> where
K: Hash + Eq + AsRef<[u8]>,

Get a mutable parameter that may or may not be a UTF-8 string.

pub fn contains_key<K: ?Sized>(&self, key: &K) -> bool where
K: Hash + Eq + AsRef<[u8]>,

Returns true if a parameter with the given key exists.

pub fn insert<K, V>(&mut self, key: K, value: V) -> Option<MaybeUtf8Owned> where
K: Into<MaybeUtf8Owned>,
V: Into<MaybeUtf8Owned>,

Insert a parameter.

pub fn remove<K: ?Sized>(&mut self, key: &K) -> Option<MaybeUtf8Owned> where
K: Hash + Eq + AsRef<[u8]>,

Remove a parameter and return it.

pub fn entry<K>(&mut self, key: K) -> Entry<MaybeUtf8Owned, MaybeUtf8Owned> where
K: Into<MaybeUtf8Owned>,

Gets the given key's corresponding parameter in the map for in-place manipulation.

pub fn parse<K: ?Sized, T>(&self, key: &K) -> Result<T, Option<T::Err>> where
K: Hash + Eq + AsRef<[u8]>,
T: FromStr,

Try to parse an entry as T, if it exists. The error will be None if the entry does not exist, and Some if it does exists, but the parsing failed.

fn my_handler(context: Context, response: Response) {
    let age: Result<u8, _> = context.variables.parse("age");
    match age {
        Ok(age) => response.send(format!("age: {}", age)),
        Err(Some(_)) => response.send("age must be a positive number"),
        Err(None) => response.send("no age provided")
    }
}

pub fn parse_or<K: ?Sized, T>(&self, key: &K, or: T) -> T where
K: Hash + Eq + AsRef<[u8]>,
T: FromStr,

Try to parse an entry as T, if it exists, or return the default in or.

fn my_handler(context: Context, response: Response) {
    let page = context.variables.parse_or("page", 0u8);
    response.send(format!("current page: {}", page));
}

pub fn parse_or_else<K: ?Sized, T, F>(&self, key: &K, or_else: F) -> T where
K: Hash + Eq + AsRef<[u8]>,
T: FromStr,
F: FnOnce(Option<T::Err>) -> T,

Try to parse an entry as T, if it exists, or create a new one using or_else. The or_else function will receive the parsing error if the value existed, but was impossible to parse.

fn my_handler(context: Context, response: Response) {
    let science = context.variables.parse_or_else("science", |_| do_heavy_stuff());
    response.send(format!("science value: {}", science));
}