Struct rustful::filter::FilterContext

pub struct FilterContext<'a> {
    pub storage: &'a mut AnyMap,
    pub log: &'a Log,
    pub global: &'a Global,
}

Contextual tools for filters.

Fields

storage

Shared storage for filters. It is local to the current request and accessible from the handler and all of the filters. It can be used to send data between these units.

log

Log for notes, errors and warnings.

global

Globally accessible data.