Struct rustful::response::Raw

pub struct Raw<'a> {
    // some fields omitted
}

A streaming fixed-size response.

Everything is written directly to the network stream, without being filtered, which makes Raw especially suitable for transferring files.

Unsafety: The content length is set beforehand, which makes it possible to send responses that are too short.

Methods

impl<'a> Raw<'a>

fn send<'d, Content: Into<Data<'d>>>(&mut self, content: Content)

Send a piece of data to the client, ignoring any eventual errors. Use try_send to get error information.

use rustful::{Context, Response};

fn my_handler(context: Context, response: Response) {
    let count = context.variables.get("count")
        .and_then(|n| n.parse().ok())
        .unwrap_or(0u8);
    let mut raw = unsafe { response.into_raw(count as u64) };

    for i in 0..count {
        raw.send([i].as_ref());
    }
}

fn try_send<'d, Content: Into<Data<'d>>>(&mut self, content: Content) -> Result<()>

Send a piece of data to the client. This is the same as send, but errors are not ignored.

use rustful::{Context, Response};

fn my_handler(context: Context, response: Response) {
    let count = context.variables.get("count")
        .and_then(|n| n.parse().ok())
        .unwrap_or(0u8);
    let mut raw = unsafe { response.into_raw(count as u64) };

    for i in 0..count {
        if let Err(e) = raw.try_send([i].as_ref()) {
            context.log.note(&format!("failed to write: {}", e));
            break;
        }
    }
}

fn end(self) -> Result<()>

Finish writing the response and collect eventual errors.

This is optional and will happen silently when the writer drops out of scope.