Struct rustful::response::Response

pub struct Response<'a, 'b> {
    // some fields omitted
}

An interface for setting HTTP status code and response headers, before data gets written to the client.

Methods

impl<'a, 'b> Response<'a, 'b>

fn new(response: Response<'a>, plugins: &'b Vec<Box<ResponsePlugin + Send + Sync>>, log: &'b Log) -> Response<'a, 'b>

fn set_status(&mut self, status: StatusCode)

Set HTTP status code. Ok (200) is default.

fn set_header<H: Header + HeaderFormat>(&mut self, header: H)

Set a HTTP response header. Date, content type (text/plain) and server is automatically set.

fn get_header<H: Header + HeaderFormat>(&self) -> Option<&H>

Get a HTTP response header if set.

fn plugin_storage(&mut self) -> &mut AnyMap

Mutably borrow the plugin storage. It can be used to communicate with the response plugins.

fn into_writer(self) -> ResponseWriter<'a, 'b>

Turn the Response into a ResponseWriter to allow the response body to be written.

Status code and headers will be written to the client and ResponsePlugin::begin() will be called on the registered response plugins.