Module rustful::router

Routers stores request handlers, using an HTTP method and a path as keys.

The TreeRouter, for example, can be created from a vector of predefined paths, like this:

#[macro_use]
extern crate rustful;
use rustful::Method::Get;
use rustful::TreeRouter;

let routes = vec![
    (Get, "/about", about_us),
    (Get, "/user/:user", show_user),
    (Get, "/product/:name", show_product),
    (Get, "/*", show_error),
    (Get, "/", show_welcome)
];

let router = TreeRouter::from_routes(routes);

Routes may also be added after the router was created, like this:

#[macro_use]
extern crate rustful;
use rustful::Method::Get;
use rustful::{Router, TreeRouter};

let mut router = TreeRouter::new();

router.insert(Get, "/about", about_us);
router.insert(Get, "/user/:user", show_user);
router.insert(Get, "/product/:name", show_product);
router.insert(Get, "/*", show_error);
router.insert(Get, "/", show_welcome);

Routers can also be used with a special macro for creating routes, called insert_routes!:

#[macro_use]
extern crate rustful;
use rustful::Method::Get;
use rustful::TreeRouter;

let router = insert_routes!{
    TreeRouter::new() => {
        "/about" => Get: about_us,
        "/user/:user" => Get: show_user,
        "/product/:name" => Get: show_product,
        "/*" => Get: show_error,
        "/" => Get: show_welcome
    }
};

This macro will generate the same code as the example above, but it allows more advanced structures to be defined without the need to write the same paths multiple times. This can be useful to lower the risk of typing errors, among other things.

Structs

TreeRouter
Stores items, such as request handlers, using an HTTP method and a path as keys.

Traits

Router
Routers are used to store request handlers.

Type Definitions

RouterResult