Module rustful::router

Routers stores request handlers, using an HTTP method and a path as keys.

Rustful provides a tree structured all-round router called TreeRouter, but any other type of router can be used, as long as it implements the Router trait. This will also make it possible to initialize it using the insert_routes! macro:

#[macro_use]
extern crate rustful;
use rustful::Method::Get;
use rustful::TreeRouter;

let router = insert_routes!{
    TreeRouter::new() => {
        "/about" => Get: about_us,
        "/users" => {
            "/" => Get: list_users,
            ":id" => Get: show_user
        },
        "/products" => {
            "/" => Get: list_products,
            ":id" => Get: show_product
        },
        "/*" => Get: show_error,
        "/" => Get: show_welcome
    }
};

This macro creates the same structure as the example below, but it allows tree structures to be defined without the need to write the same paths multiple times. This can be useful to lower the risk of typing errors, among other things.

Routes may also be added using the insert method, like this:

extern crate rustful;
use rustful::Method::Get;
use rustful::{Router, TreeRouter};

let mut router = TreeRouter::new();

router.insert(Get, "/about", about_us);
router.insert(Get, "/users", list_users);
router.insert(Get, "/users/:id", show_user);
router.insert(Get, "/products", list_products);
router.insert(Get, "/products/:id", show_product);
router.insert(Get, "/*", show_error);
router.insert(Get, "/", show_welcome);

Structs

TreeRouter
Stores items, such as request handlers, using an HTTP method and a path as keys.

Enums

RouteIter
Utility iterator for when a root path may be hard to represent.

Traits

Route
A segmented route.

Router
A common trait for routers.

Type Definitions

RouterResult