Module rustful::router

Routers stores request handlers, using an HTTP method and a path as keys.

Building Routers

Rustful provides a tree structured all-round router called TreeRouter, but any other type of router can be used, as long as it implements the Router trait. This will also make it possible to initialize it using the insert_routes! macro:

#[macro_use]
extern crate rustful;
use rustful::TreeRouter;

let router = insert_routes! {
    TreeRouter::new() => {
        Get: show_welcome,
        "about" => Get: about_us,
        "users" => {
            Get: list_users,
            ":id" => Get: show_user
        },
        "products" => {
            Get: list_products,
            ":id" => Get: show_product
        },
        "*" => Get: show_error
    }
};

This macro creates the same structure as the example below, but it allows tree structures to be defined without the need to write the same paths multiple times. This can be useful to lower the risk of typing errors, among other things.

Routes may also be added using the insert method, like this:

extern crate rustful;
use rustful::Method::Get;
use rustful::{Router, TreeRouter};

let mut router = TreeRouter::new();

router.insert(Get, "/", show_welcome);
router.insert(Get, "/about", about_us);
router.insert(Get, "/users", list_users);
router.insert(Get, "/users/:id", show_user);
router.insert(Get, "/products", list_products);
router.insert(Get, "/products/:id", show_product);
router.insert(Get, "/*", show_error);

Variables

Routes may contain variables, that are useful for capturing parts of the requested path as input to the handler. The syntax for a variable is simply an indicator character (: or *) followed by a label. Variables without labels are also valid, but their values will be discarded.

Variable Segments (:label)

A variable segment will match a single arbitrary segment. They are probably the most commonly used variables and may, for example, be used to select a blog post: "posts/:year/:month/:day/:title_slug".

pattern = "a/:v/b"
"a/c/b" -> v = "c"
"a/c/d/b" -> no match
"a/b" -> no match
"a/c/b/d" -> no match

Variable Sequences (*label)

A variable sequence is similar to a variable segment, but with the difference that it may consume multiple segments until the rest of the path gives a match. An example use case is a route for downloadable files that may be arranged in arbitrary directories: "downloads/*file_path".

pattern = "a/*v/b"
"a/c/b" -> v = "c"
"a/c/d/b" -> v = "c/d"
"a/b" -> no match
"a/c/b/d" -> no match

pattern = "a/b/*v"
"a/b/c" -> v = "c"
"a/b/c/d" -> v = "c/d"
"a/b" -> no match

Router Composition

The default tree router is actually a composition of three routers: TreeRouter, MethodRouter and Variables. They come together as the type TreeRouter<MethodRouter<Variables<_>>>, but the TreeRouter assumes that this is most probably what you want, so this is what TreeRouter::new() gives you. No need to write it all out in most of the cases.

There may, however, be cases where you want something else. What if you don't care about the HTTP method? Maybe your handler takes care of that internally. Sure, no problem:

use rustful::TreeRouter;
use rustful::router::Variables;

let my_router = TreeRouter::<Option<Variables<_>>>::default();

And what about those route variables? Not using them at all? Well, just remove them too, if you don't want them:

use rustful::TreeRouter;

let my_router = TreeRouter::<Option<_>>::default();

You can simply recombine and reorder the router types however you want, or why not make your own router? Just implement the Router trait.

Structs

Endpoint
API endpoint data.

InsertState
A state object for route insertion.

MethodRouter
A router that selects an item from an HTTP method.

RouteState
A state object for routing.

TreeRouter
A tree shaped router that selects handlers using paths.

Variables
A router endpoint that assigns names to route variables.

Enums

RouteIter
Utility iterator for when a root path may be hard to represent.

Traits

Route
A segmented route.

Router
A common trait for routers.