Module rustful::router

Router stores items, such as request handlers, using an HTTP method and a path as keys.

The Router can be created from a vector of predefined paths, like this:

let routes = [
    (Get, "/about", about_us),
    (Get, "/user/:user", show_user),
    (Get, "/product/:name", show_product),
    (Get, "/*", show_error),
    (Get, "/", show_welcome)
];

let router = Router::from_routes(&routes);

Routes may also be added after the Router was created, like this:

let mut router = Router::new();

router.insert_item(Get, "/about", about_us);
router.insert_item(Get, "/user/:user", show_user);
router.insert_item(Get, "/product/:name", show_product);
router.insert_item(Get, "/*", show_error);
router.insert_item(Get, "/", show_welcome);

There is also a special macro for creating routers, called router!(...):

#![feature(phase)]
#[phase(plugin)]
extern crate rustful_macros;

extern crate rustful;

...

let router = router!{
    "/about" => Get: about_us,
    "/user/:user" => Get: show_user,
    "/product/:name" => Get: show_product,
    "/*" => Get: show_error,
    "/" => Get: show_welcome
};

This macro will generate the same code as the example above, but it allows more advanced structures to be defined without the need to write the same paths multiple times. This can be useful to lower the risk of typing errors, among other things.

There is also a macro for creating a route vector, like the one in the first example. These can be found in the crate rustful_macros.

Structs

Router
Stores items, such as request handlers, using an HTTP method and a path as keys.

Type Definitions

RouterResult