Struct rustful::router::TreeRouter

pub struct TreeRouter<T> {
    // some fields omitted
}

Stores items, such as request handlers, using an HTTP method and a path as keys.

Paths can be static ("path/to/item") or variable ("users/:group/:user") and contain wildcards ("path/*/item/*").

Variables (starting with :) will match whatever word the request path contains at that point and it will be sent as a value to the item.

Wildcards (a single *) will consume the segments until the rest of the path gives a match.

pattern = "a/*/b"
"a/c/b" -> match
"a/c/d/b" -> match
"a/b" -> no match
"a/c/b/d" -> no match

pattern = "a/b/*"
"a/b/c" -> match
"a/b/c/d" -> match
"a/b" -> no match

Methods

impl<T> TreeRouter<T>

fn new() -> TreeRouter<T>

Creates an empty TreeRouter.

fn insert_item(&mut self, method: Method, path: &str, item: T)

Inserts an item into the TreeRouter at a given path.

impl<T: Clone> TreeRouter<T>

fn from_routes(routes: &[(Method, &str, T)]) -> TreeRouter<T>

Generates a TreeRouter tree from a set of items and paths.

fn insert_router(&mut self, path: &str, router: &TreeRouter<T>)

Insert an other TreeRouter at a path. The content of the other TreeRouter will be merged with this one. Content with the same path and method will be overwritten.