Trait rustful::router::Router

pub trait Router {
    type Handler;

    fn find(&self, method: &Method, path: &str) -> Option<(&Self::Handler, HashMap<String, String>)>;
    fn insert(&mut self, method: Method, path: &str, handler: Self::Handler);
}

Routers are used to store request handlers.

A router implementing the Router trait will always be compatible with the insert_routes! macro.

Associated Types

type Handler

Required Methods

fn find(&self, method: &Method, path: &str) -> Option<(&Self::Handler, HashMap<String, String>)>

Find and return the matching handler and variable values.

fn insert(&mut self, method: Method, path: &str, handler: Self::Handler)

Insert a new handler into the router.