Trait rustful::router::Router

pub trait Router: Send + Sync + 'static {
    type Handler: Handler;
    fn insert<'a, D: ?Sized + Deref<Target=R> + 'a, R: ?Sized + Route<'a> + 'a>(&mut self, method: Method, route: &'a D, handler: Self::Handler);
    fn find<'a>(&'a self, method: &Method, route: &[u8]) -> Endpoint<'a, Self::Handler>;
}

A common trait for routers.

A router must to implement this trait to be usable in a Rustful server. This trait will also make the router compatible with the insert_routes! macro.

Associated Types

type Handler: Handler

The request handler type that is stored within this router.

Required Methods

fn insert<'a, D: ?Sized + Deref<Target=R> + 'a, R: ?Sized + Route<'a> + 'a>(&mut self, method: Method, route: &'a D, handler: Self::Handler)

Insert a new handler into the router.

fn find<'a>(&'a self, method: &Method, route: &[u8]) -> Endpoint<'a, Self::Handler>

Find and return the matching handler and variable values.