Trait rustful::router::Router

pub trait Router {
    type Handler;

    fn insert<'r, R: Route<'r> + ?Sized>(&mut self, method: Method, route: &'r R, handler: Self::Handler);
    fn find<'a, 'r, R: Route<'r> + ?Sized>(&'a self, method: &Method, route: &'r R) -> RouterResult<'a, Self::Handler>;
}

A common trait for routers.

A router must to implement this trait to be usable in a Rustful server. This trait will also make the router compatible with the insert_routes! macro.

Associated Types

type Handler

Required Methods

fn insert<'r, R: Route<'r> + ?Sized>(&mut self, method: Method, route: &'r R, handler: Self::Handler)

Insert a new handler into the router.

fn find<'a, 'r, R: Route<'r> + ?Sized>(&'a self, method: &Method, route: &'r R) -> RouterResult<'a, Self::Handler>

Find and return the matching handler and variable values.