Struct rustful::server::Server

pub struct Server<R: Router> {
    pub handlers: R,
    pub fallback_handler: Option<R>,
    pub host: Host,
    pub scheme: Scheme,
    pub threads: Option<usize>,
    pub server: String,
    pub content_type: Mime,
    pub log: Box<Log>,
    pub global: Global,
    pub context_filters: Vec<Box<ContextFilter>>,
    pub response_filters: Vec<Box<ResponseFilter>>,
}

Used to set up and run a server.

let server_result = Server {
    host: 8080.into(),
    handlers: router,
    ..Server::default()
}.run();

match server_result {
    Ok(_server) => {},
    Err(e) => println!("could not start server: {}", e.description())
}

Fields

handlers

One or several response handlers.

fallback_handler

A fallback handler for when none is found in handlers. Leaving this unspecified will cause an empty 404 response to be automatically sent instead.

host

The host address and port where the server will listen for requests. Default is 0.0.0.0:80.

scheme

Use good old HTTP or the more secure HTTPS. Default is HTTP.

threads

The number of threads to be used in the server thread pool. The default (None) will cause the server to use a value based on recommendations from the system.

server

The content of the server header. Default is "rustful".

content_type

The default media type. Default is text/plain, charset: UTF-8.

log

Tool for printing to a log. The default is to print to standard output.

global

Globally accessible data.

context_filters

The context filter stack.

response_filters

The response filter stack.

Methods

impl<R: Router> Server<R>

fn new(handlers: R) -> Server<R>

Set up a new standard server. This can be useful when handlers doesn't implement Default:

let handler = |context: Context, response: Response| {
    //...
};

let server_result = Server {
    host: 8080.into(),
    ..Server::new(handler)
};

fn run(self) -> HttpResult<Listening>

Start the server.

fn build(self) -> (ServerInstance<R>, Scheme)

Build a runnable instance of the server.