Struct rustful::server::Server

pub struct Server<'s, R> {
    pub handlers: R,
    pub host: SocketAddr,
    pub protocol: Scheme<'s>,
    pub threads: Option<usize>,
    pub server: String,
    pub content_type: Mime,
    pub log: Box<Log + Send + Sync>,
    pub context_plugins: Vec<Box<ContextPlugin + Send + Sync>>,
    pub response_plugins: Vec<Box<ResponsePlugin + Send + Sync>>,
}

Used to build and run a server.

Each field has a corresponding modifier method and calls to these methods can be chained for quick setup.

let server_result = Server::new().port(8080).handlers(router).run();

match server_result {
    Ok(_server) => {},
    Err(e) => println!("could not start server: {}", e.description())
}

Fields

handlers

One or several response handlers.

host

The host address and port where the server will listen for requests.

protocol

The type of HTTP protocol.

threads

The number of threads to be used in the server task pool. The default (None) is based on recommendations from the system.

server

The content of the server header.

content_type

The default media type.

log

Tool for printing to a log.

context_plugins

The context plugin stack.

response_plugins

The response plugin stack.

Methods

impl<'s> Server<'s, ()>

fn new() -> Server<'s, ()>

Create a new empty server which will listen on host address 0.0.0.0 and port 80.

impl<'s, R> Server<'s, R>

fn handlers<NewRouter: Router<Handler=H>, H: Handler>(self, handlers: NewRouter) -> Server<'s, NewRouter>

Set request handlers.

fn port(self, port: u16) -> Server<'s, R>

Change the port. Default is 80.

fn host(self, host: SocketAddr) -> Server<'s, R>

Change the host address. Default is 0.0.0.0:80.

fn https(self, cert: &'s Path, key: &'s Path) -> Server<'s, R>

Change the protocol to HTTPS.

fn threads(self, threads: Option<usize>) -> Server<'s, R>

Set the number of threads to be used in the server task pool.

Passing None will set it to the default number of threads, based on recommendations from the system.

fn server_name(self, name: String) -> Server<'s, R>

Change the server response header. Default is rustful.

fn content_type(self, content_type: Mime) -> Server<'s, R>

Change the default content type. Default is text/plain.

fn log<L: Log + Send + Sync + 'static>(self, log: L) -> Server<'s, R>

Change log tool. Default is to print to standard output.

fn with_context_plugin<P: ContextPlugin + Send + Sync + 'static>(self, plugin: P) -> Server<'s, R>

Add a context plugin to the plugin stack.

fn with_response_plugin<P: ResponsePlugin + Send + Sync + 'static>(self, plugin: P) -> Server<'s, R>

Add a response plugin to the plugin stack.

impl<'s, R, H> Server<'s, R> where R: Router<Handler=H> + Send + Sync + 'static, H: Handler + Send + Sync + 'static

fn run(self) -> HttpResult<Listening>

Start the server.

fn build(self) -> (ServerInstance<R>, Scheme<'s>)

Build a runnable instance of the server.