Struct rustful::server::ServerInstance

pub struct ServerInstance<R> { /* fields omitted */ }

A runnable instance of a server.

It's not meant to be used directly, unless additional control is required.

let server_instance = Server {
    host: 8080.into(),
    handlers: router,
    ..Server::default()
}.build();

Methods

impl<R: HandleRequest + 'static> ServerInstance<R>

pub fn new(config: Server<R>) -> ServerInstance<R>

Create a new server instance, with the provided configuration. This is the same as Server{...}.build().

pub fn run(self) -> HttpResult<Listening>

Start the server.

pub fn run_https<S: SslServer + Clone + Send + 'static>(
self,
ssl: S
) -> HttpResult<Listening>

Start the server with SSL.