Struct rustful::Server

pub struct Server<H, C> {
    // some fields omitted
}

Receives the HTTP requests and passes them on to handlers.

let server = Server::new(8080, Router::from_routes(routes));

server.run();

Methods

impl<H: Handler<()> + Send + Sync> Server<H, ()>

fn new(port: Port, handlers: Router<H>) -> Server<H, ()>

Create a new Server which will listen on the provided port and host address 0.0.0.0.

impl<H: Handler<C> + Send + Sync, C: Cache + Send + Sync> Server<H, C>

fn with_cache(port: Port, cache: C, handlers: Router<H>) -> Server<H, C>

Creates a new Server with a resource cache.

The server will listen listen on the provided port and host address 0.0.0.0. Cache cleaning is disabled by default.

fn run(self)

Start the server and run forever. This will only return if the initial connection fails.

impl<H, C> Server<H, C>

fn set_host(&mut self, host: IpAddr)

Change the host address.

fn set_clean_interval(&mut self, interval: Option<u32>)

Set the minimal number of seconds between each cache clean.

Passing None disables cache cleaning.

fn set_server_name(&mut self, name: String)

Change the server response header.

fn set_content_type(&mut self, content_type: MediaType)

Change the default content type.

fn add_request_plugin<P: RequestPlugin + Send + Sync>(&mut self, plugin: P)

Add a request plugin to the plugin stack.

fn add_response_plugin<P: ResponsePlugin + Send + Sync>(&mut self, plugin: P)

Add a response plugin to the plugin stack.