Struct rustful::TreeRouter

pub struct TreeRouter<T: Router + Default> {
    pub find_hyperlinks: bool,
    // some fields omitted
}

A tree shaped router that selects handlers using paths.

Each tree node stores an other router of type T, which has to implement Default to allow "empty" nodes, and a number of child references. The requested path must always be exhausted before sub-routers are searched, so there is no point in storing other path based routers in a TreeRouter.

The TreeRouter has support for shallow hyperlinks to children, siblings, cousins, and so forth. The use of variable sequences complicates this process and may cause confusing results in certain situations. The hyperlinks may or may not point to a handler.

Hyperlinks has to be activated by setting find_hyperlinks to true.

Fields

find_hyperlinks

Should the router search for hyperlinks? Setting this to true may slow down endpoint search, but enables hyperlinks.

Methods

impl<H: Handler> TreeRouter<MethodRouter<Variables<H>>>

fn new() -> TreeRouter<MethodRouter<Variables<H>>>

Creates an empty TreeRouter<MethodRouter<Variables<H>>>, which is probably the most common composition. It will select handlers based on path and then HTTP method, and collect any variables on the way.

Use TreeRouter::default() if this is not the desired composition.